Remove all easily removable nqp::stmts
Basically, whenever an nqp::stmts aligned with an outer scope, and
it was already laid out in a familiar way (to me at least).  Another
step toward de-nqpifying the core setting.

Should not have any semantic changes, nor noticeable performance
effects.  Should allow for better error reporting (because an execution
error inside an nqp::stmts is reported at the start of the nqp::stmts,
which can be a long way off from the actual location of the error).