Change nqp::tostr_I(foo) to faster nqp::base_I(foo,16)
As just discussed on #moarvm, as it is orders of magnitude faster.
This most notably affects Int.WHICH, which will now report its
value in hex: before 42.WHICH was Int|42, now it is Int|2A

Note that this will only show up as an improvement in certain
workloads, that depend heavily on WHICH values, such as using
object hashes and set operations.