Revert "Stop REPL complaining about "useless use""
This reverts commit eae309a49f5e4274d47e966b6f3fcbf735ce0803.

This is opening another nest of bugs, so closing this for now.

Sadly, the REPL in its current state has a CONTROL block to catch
CONTROL messages such as warnings, but sadly this appears to be
never called.  It looks like warnings are actually captured in NQP
and thus not interceptable in Raku  :-(