Properly fix improper use of () in p6bindattrvinres
() is *NOT* a new instance of the empty List, it *is* the empty List
and should as such not be used to put things into.  The reason this
wasn't spotted before, is probably because this functionality is not
used an awful load in most use cases.  Spotted by Stefan Seifert++